Restore the ability to call the module's entry point.
I had read that it is considered an anti-pattern to create package scripts inside the package itself, but now I can't find that source or the reasoning for it. So I moved the file to the top-level directory (the same one that holds pyproject.toml), but then I couldn't get [project.scripts] working correctly for the entry point. Thus, I will stick with the pattern used in my other recent packages.